+++ firmware/CHANGELOG.md
@@ -31,6 +31,8 @@ All notable user-facing or behavior-altering changes will be documented in this 
  - Ford PIP trigger decoder https://github.com/rusefi/rusefi/wiki/Images/triggers/trigger_65.png
  - Dwell battery voltage correction table
  - rusEFI firmware feature should be honest about it's Windows-only limitation
+ - Bosch MM5.10 accelerometer integration
+ - CLT/IAT sensor calibration presets
 
 ### Fixed
  - Fix of 36/2/2/2 attributes to help rotary #3242